requirements.txt
Flask==3.1.0
urllib3==2.3.0
werkzeug==3.1.3
pandas
scikit-learn
joblib
numpy1
gunicorn==21.2.0
validators
requests
beautifulsoup4
html2image
